fix: push notification

+++ comments/receivers.py
@@ -3,6 +3,7 @@ from django.db.models.signals import post_save, post_delete
 
 from accounts.models import Fcmtoken
 from exercises.models import Exercise
+from notifications.serializers import NotificationSerializer
 from .models import Comment
 import os
 from pathlib import Path
@@ -14,11 +15,14 @@ import environ
 env = environ.Env()
 
 from django.contrib.auth import get_user_model
+from django.contrib.humanize.templatetags import humanize
 
 User = get_user_model()
 
 import firebase_admin
 from firebase_admin import credentials, messaging
+
+import json
 
 BASE_DIR = Path(__file__).resolve().parent.parent
 
@@ -31,42 +35,60 @@ firebase_admin.initialize_app(cred)
 
 @receiver(post_save, sender=Comment)
 def create_notification(sender, instance, created, **kwargs):
+
     if created:
         user = instance.user
         name = user.get_full_name()
         exercise = Exercise.objects.get(id=instance.object_id)
         content = ContentType.objects.get_for_model(Comment)
 
-        Notification.objects.create(
+        notification = Notification.objects.create(
             user=user,
             user_to=exercise.user,
             content_type=content,
             object_id=instance.id,
             # link=f"/exercises/{exercise.id}/comments",
         )
+        serializer = NotificationSerializer(notification)
 
         data = Fcmtoken.objects.filter(user=exercise.user)
 
         for item in data:
-            sendNotification(item.token, name, exercise, instance)
+            sendNotification(
+                item.token, notification, user, name, exercise, instance, serializer
+            )
 
 
-def sendNotification(fcm_token, name, exercise, instance):
-    frontendUrl = env("FRONTEND_URL")
+def sendNotification(
+    fcm_token, notification, user, name, exercise, instance, serializer
+):
+    backend_url = env("BACKEND_URL")
+    frontend_url = env("FRONTEND_URL")
+
     options = messaging.WebpushFCMOptions(
-        link=frontendUrl + "/workouts",
+        link=frontend_url + "/workouts",
     )
+
+    # created_format = humanize.naturaltime(instance.created)
+
+    # only string data
+    json_data = {
+        "id": str(notification.id),
+        "username": user.username,
+        "photo": backend_url + user.photo.url,
+        "comment": instance.content,
+        "exercise": exercise.name,
+        "created": str(instance.created),
+        "is_read": str(False),
+    }
+
     message = messaging.Message(
         notification=messaging.Notification(
             title="AFit Trainer",
             body=f"{name} comento sobre el ejercicio {exercise.name}: {instance.content}",
         ),
         webpush=messaging.WebpushConfig(fcm_options=options),
-        data={
-            "link": frontendUrl + "/workouts",
-            "title": "AFit Trainer",
-            "body": f"{name} comento sobre el ejercicio {exercise.name}: {instance.content}",
-        },
+        data=json_data,
         token=fcm_token,
     )
 
@@ -75,5 +97,9 @@ def sendNotification(fcm_token, name, exercise, instance):
 
 @receiver(post_delete, sender=Comment)
 def deleteNotification(sender, instance, **kwargs):
-    notification = Notification.objects.get(pk=instance.id)
-    notification.delete()
+    try:
+        notification = Notification.objects.get(object_id=instance.id)
+        if notification:
+            notification.delete()
+    except Notification.DoesNotExist:
+        return
